+++ CHANGES.md
@@ -62,7 +62,9 @@ Dashboad CLientes
         M[site.php]
         M[pages/site_sidebar.php]
         M[pages/cdc_dashboard.php]
-
+    - Agrego DB de totales: servicios, cpu, ram, storage, VMs, Clientes
+        M[pages/cdc_dashboard.php]
+        
 ## FEAT-CLI-DC2
 
 Correcciones y mejoras POST Producción